refactor(activesupport): split time-ext.ts onto the Rails file layout, converge TimeZone#match? and dst?
RFC 0098. Two of the three bundled stories; the third is blocked.

Split by receiver (split-time-ext-by-receiver-onto-the-rails-layout).
`time-ext.ts` hosted members from four Rails files on three receivers, so
`parity:api` — which matches on the Rails file layout — could only credit one
receiver per name. Bodies travel unchanged into the paths `conventions.ts`
derives from their Ruby counterparts:

- `core_ext/time/conversions.rb` → `core-ext/time/conversions.ts`
  (`DATE_FORMATS`, `to_fs`/`to_formatted_s`, `formatted_offset`)
- `core_ext/time/compatibility.rb` → `core-ext/time/compatibility.ts`
  (`to_time`, `preserve_timezone`, `system_local_time?`,
  `active_support_local_zone`); `core_ext/date_time/compatibility.rb`'s
  `to_time` is the other arm of the same function and buckets there too
- `core_ext/string/zones.rb` → `core-ext/string/zones.ts` (`in_time_zone`)

`Date#readable_inspect` / `#default_inspect` (`date/conversions.rb:63-66`) and
`Date#xmlschema` (`:95-97`, `in_time_zone.xmlschema`) become writable now that
the receivers no longer share a file, and land on `core-ext/date/conversions.ts`
next to that file's `to_fs`. The `time-ext.ts` call-mismatch row the old
`xmlschema` carried converges away and is deleted.

AR closure: 8917/8943 → 8931/8943.

TimeZone#match? / #dst? (converge-time-zone-match-and-dst-predicates).
`match(identifier)` compared `name` and the resolved `tzinfo.identifier` and
dropped Rails' `Regexp` arm entirely (`values/time_zone.rb:348-351`). It is now
`isMatch(re)` with all three arms — `re === name`, `re === MAPPING[name]`, and
the `RegExp` arm testing both — and `TimeZoneTest`'s two cases assert what
Rails' `test_zone_match` / `test_zone_match?` do. `isDst`'s invented
`Temporal.Now.instant()` default is gone: `dst?(time)` (`:571-573`) takes a
required argument.

`port-parts-to-time-behind-time-at` is blocked: `parts_to_time` needs
`Time.at`, which `@blazetrails/date` does not have, and Rails reaches it only
from `parse`/`strptime`, which trails implements as bespoke component scanners
that never build a parts hash — a PR of its own.

Closes-story: converge-time-zone-match-and-dst-predicates
Closes-story: split-time-ext-by-receiver-onto-the-rails-layout
