+++ packages/activesupport/src/core-ext/date/conversions.ts
@@ -11,6 +11,7 @@ import { ordinalize } from "../../inflector.js";
 import { TimeWithZone } from "../../time-with-zone.js";
 import { TimeZone } from "../../values/time-zone.js";
 import { ArgumentError } from "../../time-zone-config.js";
+import { inTimeZone } from "../date-and-time/zones.js";
 
 /**
  * The named formats `Date#to_fs` resolves, either a `strftime` string or a
@@ -65,6 +66,29 @@ export function toFs(date: Temporal.PlainDate, format: string = "default"): stri
 export { toFs as toFormattedS };
 
 /**
+ * Overrides the default inspect method with a human readable one, e.g.,
+ * "Mon, 21 Feb 2005".
+ *
+ * Mirrors: `Date#readable_inspect` (`core_ext/date/conversions.rb:63-65`) —
+ * `strftime("%a, %d %b %Y")`. Rails then runs
+ * `alias_method :inspect, :readable_inspect`, reopening `::Date` itself;
+ * trails has no receiver to reopen, so callers reach this name directly.
+ */
+export function readableInspect(date: Temporal.PlainDate): string {
+  return new RubyDate(date).strftime("%a, %d %b %Y");
+}
+
+/**
+ * Mirrors: `alias_method :default_inspect, :inspect`
+ * (`core_ext/date/conversions.rb:66`), which captures the ORIGINAL
+ * `::Date#inspect` — ruby/date's `d_lite_inspect` — before the line below it
+ * replaces `inspect` with {@link readableInspect}.
+ */
+export function defaultInspect(date: Temporal.PlainDate): string {
+  return new RubyDate(date).inspect();
+}
+
+/**
  * Mirrors: `Date#to_time` (`core_ext/date/conversions.rb:83-86`) —
  * `::Time.public_send(form, year, month, day)`.
  *
@@ -86,4 +110,19 @@ export function toTime(date: Temporal.PlainDate, form: string = "local"): TimeWi
   }
   const zone = TimeZone.find(form === "utc" ? "UTC" : Temporal.Now.timeZoneId())!;
   return zone.local(date.year, date.month, date.day);
+}
+
+/**
+ * Returns a string which represents the time in used time zone as DateTime
+ * defined by XML Schema:
+ *
+ *     date = Date.new(2015, 5, 23)  // => Sat, 23 May 2015
+ *     xmlschema(date)               // => "2015-05-23T00:00:00+04:00"
+ *
+ * Mirrors: `Date#xmlschema` (`core_ext/date/conversions.rb:95-97`) —
+ * `in_time_zone.xmlschema`, so the offset is the one `Time.zone` (or the
+ * system zone) puts the day's midnight at, not a bare `Z`.
+ */
+export function xmlschema(date: Temporal.PlainDate): string {
+  return inTimeZone(date).xmlschema();
 }
